requirements.txt
Pillow==9.5.0
segno==1.5.2
textwrap3==0.9.2
